@@ -31,7 +31,7 @@ import io.javaoperatorsdk.operator.api.reconciler.Context;
 import org.junit.jupiter.api.BeforeEach;
 
 /**
- * @link JobStatusObserver unit tests
+ * @link Base class for unit tests
  */
 public abstract class OperatorTestBase {
 
@@ -70,7 +70,21 @@ public abstract class OperatorTestBase {
     }
 
     public <CR extends AbstractFlinkResource<?, ?>> FlinkResourceContext<CR> getResourceContext(
+            CR cr, Configuration configuration) {
+        return getResourceContext(cr, context, configuration);
+    }
+
+    public <CR extends AbstractFlinkResource<?, ?>> FlinkResourceContext<CR> getResourceContext(
             CR cr, Context josdkContext) {
+        var ctxFactory =
+                new TestingFlinkResourceContextFactory(
+                        configManager, operatorMetricGroup, flinkService, eventRecorder);
+        return ctxFactory.getResourceContext(cr, josdkContext);
+    }
+
+    public <CR extends AbstractFlinkResource<?, ?>> FlinkResourceContext<CR> getResourceContext(
+            CR cr, Context josdkContext, Configuration configuration) {
+        configManager.updateDefaultConfig(configuration);
         var ctxFactory =
                 new TestingFlinkResourceContextFactory(
                         configManager, operatorMetricGroup, flinkService, eventRecorder);
